\name{org-package}
\alias{org-package}
\alias{org}
\docType{package}
\title{
  Org
}
\description{
  Helper functions to interact with org files: read org tables, convert
  data.frames to org tables.
}
\details{

  Helper functions to interact with org files: read org tables, convert
  data.frames to org tables.
  
}
\author{
  Enrico Schumann <es@enricoschumann.net>
}
\references{
  Org mode manual
}
\keyword{ package }
\seealso{
  \code{\link{toOrg}}
}
