\name{orgutils-package}
\alias{orgutils-package}
\alias{org}
\docType{package}
\title{
  Org Utils
}
\description{
  Helper functions to interact with org files: read org tables, convert
  R objects to org markup.
}
\details{

  Helper functions to interact with org files: read org tables, convert
  R objects to org markup.

  There are several other packages that help you work with org files as
  well, such as \pkg{orgR}.
  
}
\author{
  Enrico Schumann <es@enricoschumann.net>
}
\references{
  Org mode manual \url{http://orgmode.org/}
}
\keyword{ package }
\seealso{
  \code{\link{toOrg}}, \code{\link{readOrg}}
}
